Partially finished update
Not many changes or additions, but enough to warrant paraphrasing and skipping most of them.

I will finish the rest of the soon, I am just on the verge of passing out due to sleep deprivation.

Added an event system, hopefully no one runs into any issues using it as it's fairly simple and shouldn't really be complicated enough to not work as intended. :^)

Added the classes for all the modules I intend to add in this base. Will add more utils to the client at some point too.

Suk Yer Mam - Shae